Add tags to the global enum type during component initialization
This creates tags as Module::Name. For example, you would have Analyzer::HTTP
in a global Tag enum type. This has the same enum value as Analyzer::ANALYZER_HTTP.
It also removes the need to pass an EnumType down to the Tag and Component
constructors. This causes some changes in test output, mostly around tag naming.